refactor(portal): feature-folder settings tabs, delete dead panels
Each substantial settings tab now owns one feature folder under features/,
with a single name and a public index.ts barrel; components/settings/ keeps
only the shell (layout, sidebar, runtime-provider).

- launch: consolidate the deploy UI (was components/settings/onboarding/, the
  confusing second name) into features/launch/components/. The logic barrel
  (features/launch/index.ts) stays logic-only; components/index.ts is the UI
  entry, so there is no import cycle. settings-layout mounts
  @portal/features/launch/components.
- general/apps/app-keys/bots/secrets/byok: moved from components/settings/*.tsx
  into features/<tab>/ with an index.ts barrel.
- settings-styles.ts (pure className constants, shared) moved to lib/.
- Delete 5 dead panels with zero importers: account-settings, privacy-settings,
  billing-settings, capabilities-settings, connectors-settings.

Internal structure is right-sized: heavy tabs keep their state/lifecycle in
place; the convention (one folder + barrel per tab, server/client separation
unchanged, enforced by the existing no-restricted-imports rule) is uniform.

Verified: tsc, eslint, vitest (165), next build all green.
