--- apps/portal/src/components/settings/billing-settings.tsx
@@ -1,14 +1,0 @@
-"use client";
-
-export function BillingSettings() {
-  return (
-    <div className="space-y-8">
-      <div>
-        <h3 className="text-lg font-semibold text-gray-900 mb-4">Billing</h3>
-        <p className="text-sm text-gray-500">
-          Manage your subscription and billing information.
-        </p>
-      </div>
-    </div>
-  );
-}
