--- apps/portal/src/components/settings/capabilities-settings.tsx
@@ -1,14 +1,0 @@
-"use client";
-
-export function CapabilitiesSettings() {
-  return (
-    <div className="space-y-8">
-      <div>
-        <h3 className="text-lg font-semibold text-gray-900 mb-4">Model Capabilities</h3>
-        <p className="text-sm text-gray-500">
-          Configure which capabilities are available for your conversations.
-        </p>
-      </div>
-    </div>
-  );
-}
