--- apps/portal/src/components/settings/connectors-settings.tsx
@@ -1,14 +1,0 @@
-"use client";
-
-export function ConnectorsSettings() {
-  return (
-    <div className="space-y-8">
-      <div>
-        <h3 className="text-lg font-semibold text-gray-900 mb-4">Connectors</h3>
-        <p className="text-sm text-gray-500">
-          Connect external services and tools to enhance your conversations.
-        </p>
-      </div>
-    </div>
-  );
-}
