+++ apps/portal/src/features/app-keys/app-keys.tsx
@@ -18,7 +18,7 @@ import {
   settingsSubTitleClass,
   settingsTableCardClass,
   settingsTitleClass,
-} from "./settings-styles";
+} from "@portal/lib/settings-styles";
 
 type OwnedAppKey = {
   key_hash: string;
