+++ apps/portal/src/features/app-keys/index.ts
@@ -1,0 +1,1 @@
+export { AppKeys } from "./app-keys";
