+++ apps/portal/src/features/apps/apps-settings.tsx
@@ -14,7 +14,7 @@ import {
   settingsSubTitleClass,
   settingsTableCardClass,
   settingsTitleClass,
-} from "./settings-styles";
+} from "@portal/lib/settings-styles";
 
 type AppRow = {
   app: string;
