+++ apps/portal/src/features/apps/index.ts
@@ -1,0 +1,1 @@
+export { AppsSettings } from "./apps-settings";
