+++ apps/portal/src/features/bots/bots.tsx
@@ -17,7 +17,7 @@ import {
   settingsSubTitleClass,
   settingsTableCardClass,
   settingsTitleClass,
-} from "./settings-styles";
+} from "@portal/lib/settings-styles";
 
 type BotRegistration = {
   id: string;
