+++ apps/portal/src/features/bots/index.ts
@@ -1,0 +1,1 @@
+export { Bots } from "./bots";
