+++ apps/portal/src/features/byok/byok.tsx
@@ -17,7 +17,7 @@ import {
   settingsSubTitleClass,
   settingsTableCardClass,
   settingsTitleClass,
-} from "./settings-styles";
+} from "@portal/lib/settings-styles";
 
 const PROVIDERS = [
   { id: "openai", label: "OpenAI" },
