+++ apps/portal/src/features/general/index.ts
@@ -1,0 +1,1 @@
+export { GeneralSettings } from "./general-settings";
