+++ apps/portal/src/features/launch/components/index.ts
@@ -1,0 +1,4 @@
+// Public UI surface of the launch feature. The logic barrel is `../index`
+// (state/client/contracts/…); components import that, never this file, so
+// there is no import cycle.
+export { DeployDashboard } from "./deploy-dashboard";
