+++ apps/portal/src/features/secrets/index.ts
@@ -1,0 +1,1 @@
+export { Secrets } from "./secrets";
